requirements.txt
# Streamlit 기반 UI
streamlit==1.31.1

# LangChain 핵심 프레임워크
langchain==0.1.16
langchain-openai==0.1.3
langchain-community==0.0.34
langchain-core==0.1.47

# 벡터 DB (FAISS)
faiss-cpu==1.8.0

# Azure OpenAI 사용을 위한 openai SDK
openai==1.23.2

# PDF 문서 로더
pypdf==4.1.0

# DuckDuckGo 웹 검색
duckduckgo-search==4.5

# 기타 유틸리티
python-dotenv==1.0.1
